Move main code to CommandLineTool
